Added proper commentaries to my code and split the classes from 'Universe.cs' into multiple files
